User class now accepts World instance
Previously, the User class called World.getWorld() to get the singleton
World instance. Now, the User class accepts a World instance to use.
In the case of SSH authenticaton, the World is originally passed as the
sole parameter to SSHFactoryFactory(). This is now passed to the SSHRealm,
which saves it and uses it when creating SSHUser instances. SSHUser is
itself a subclass of User.